move SchemaManager and value_is_empty
