+++ tdrs-backend/tdpservice/parsers/schema_defs/tanf/t1.py
@@ -1,8 +1,7 @@
 """Schema for t1 record types."""
 
-from ...util import SchemaManager
 from ...fields import Field
-from ...row_schema import RowSchema
+from ...row_schema import SchemaManager, RowSchema
 from ... import validators
 from tdpservice.search_indexes.models.tanf import TANF_T1
