+++ tdrs-backend/tdpservice/parsers/schema_defs/tanf/t2.py
@@ -1,10 +1,9 @@
 """Schema for HEADER row of all submission types."""
 
 
-from ...util import SchemaManager
 from ...transforms import tanf_ssn_decryption_func
 from ...fields import TransformField, Field
-from ...row_schema import RowSchema
+from ...row_schema import SchemaManager, RowSchema
 from ... import validators
 from tdpservice.search_indexes.models.tanf import TANF_T2
