+++ tdrs-backend/tdpservice/parsers/schema_defs/tanf/t4.py
@@ -1,9 +1,8 @@
 """Schema for HEADER row of all submission types."""
 
 
-from ...util import SchemaManager
 from ...fields import Field
-from ...row_schema import RowSchema
+from ...row_schema import SchemaManager, RowSchema
 from ... import validators
 from tdpservice.search_indexes.models.tanf import TANF_T4
